requirements.txt
aiosqlite>=0.19.0  # Async SQLite database operations
asyncio>=3.4.3    # Asynchronous I/O
scapy>=2.5.0     # Network packet manipulation
ipwhois>=1.2.0   # IP WHOIS lookup functionality
geopy>=2.3.0     # Geographical IP location
aiosmtplib>=2.0.1 # Async SMTP for email notifications
nvdlib>=0.7.3    # NVD/CVE database access
aiohttp>=3.8.4   # Async HTTP for CVE database updates
schedule>=1.1.0   # Task scheduling
pyyaml>=6.0      # YAML configuration support
requests>=2.31.0  # HTTP requests for ExploitDB
beautifulsoup4>=4.12.0  # HTML parsing for ExploitDB
gitpython>=3.1.40  # Git operations for ExploitDB
